Release 10.6.0 (#1142)
* Always log command even at ERROR level

* Fix help message for jobs list

* Updated sync to support the 2 smb flags

* Jumbo Blob support (#1113)

* Support Jumbo Blobs

* Jumbo blobs: Improve chunking logic

* Improve chunking logic

* Review Ze's comments

* Add test to verify block size parameters

* Improve error messaging and STE data schema version

* Support download/delete of blob versions (#1123)

* Changes in Front-end side

* Changes for download (single version working fine)

* Minor Fix

* Fix | first review

* Fix | Breaking UTs

* Fix | moved list-of-versions flag processing in goroutine

* Checked remove flow | fix minor issues

* Removed hack

* Check to ensure source==blob and destination==localdirectory

* Review Minor Fixes

* Refactoring

* Issues Fixed shown by StgExp

* Release V10.6.0 Changelogs (#1141)

* Release V10.6.0 Changelogs

* Updated message

+++ ste/JobsAdmin.go
@@ -149,7 +149,7 @@ func initJobsAdmin(appCtx context.Context, concurrency ConcurrencySettings, targ
 	if targetRateInMegaBitsPerSec > 0 {
 		// use the "networking mega" (based on powers of 10, not powers of 2, since that's what mega means in networking context)
 		targetRateInBytesPerSec := int64(targetRateInMegaBitsPerSec * 1000 * 1000 / 8)
-		unusedExpectedCoarseRequestByteCount := uint32(0)
+		unusedExpectedCoarseRequestByteCount := int64(0)
 		pacer = newTokenBucketPacer(targetRateInBytesPerSec, unusedExpectedCoarseRequestByteCount)
 		// Note: as at July 2019, we don't currently have a shutdown method/event on JobsAdmin where this pacer
 		// could be shut down. But, it's global anyway, so we just leave it running until application exit.
